Make DummyHwnd public and move Fence to DX12
